+++ tests/bunit.core.tests/TestContextBaseTest.net5.cs
@@ -1,9 +1,10 @@
-
 #if NET5_0_OR_GREATER
+
+using Bunit.TestAssets.SampleComponents.DisposeComponents;
 
 namespace Bunit;
 
-public class TestContextBaseTest : TestContext
+public partial class TestContextBaseTest : TestContext
 {
 	[Fact(DisplayName = "ComponentFactories CanCreate() method are checked during component instantiation")]
 	public void Test0001()
@@ -45,6 +46,73 @@ public class TestContextBaseTest : TestContext
 		secondMock.Verify(x => x.Create(typeof(Simple1)), Times.Once);
 	}
 
+	[Fact(DisplayName = "DisposeComponents captures exceptions from DisposeAsync in Renderer.UnhandledException")]
+	public async Task Test201()
+	{
+		RenderComponent<AsyncThrowExceptionComponent>();
+
+		DisposeComponents();
+
+		var exception = await Renderer.UnhandledException;
+		exception.ShouldBeOfType<NotSupportedException>();
+	}
+
+	[Fact(DisplayName = "DisposeComponents calls DisposeAsync on rendered components")]
+	public async Task Test202()
+	{
+		var cut = RenderComponent<AsyncDisposableComponent>();
+		var instance = cut.Instance;
+
+		DisposeComponents();
+
+        // Windows timer resolution is around 15ms therefore we want to have a higher value than the test
+        // itself to prohibit flakiness
+		await Task.Delay(25);
+		instance.WasDisposed.ShouldBeTrue();
+	}
+
+	[Fact(DisplayName = "DisposeComponents should dispose components added via ComponentFactory")]
+	public void Test203()
+	{
+		ComponentFactories.Add<ChildDispose, MyChildDisposeStub>();
+		var cut = RenderComponent<ParentDispose>(ps => ps.Add(p => p.CallStack, new List<string>()));
+		var instance = cut.FindComponent<MyChildDisposeStub>().Instance;
+
+		DisposeComponents();
+
+		instance.WasDisposed.ShouldBeTrue();
+	}
+
+	private sealed class MyChildDisposeStub : ComponentBase, IDisposable
+	{
+		public bool WasDisposed { get; private set; }
+
+		public void Dispose()
+		{
+			WasDisposed = true;
+		}
+	}
+
+	private sealed class AsyncThrowExceptionComponent : ComponentBase, IAsyncDisposable
+	{
+		public async ValueTask DisposeAsync()
+		{
+			await Task.Delay(10);
+			throw new NotSupportedException();
+		}
+	}
+
+	private sealed class AsyncDisposableComponent : ComponentBase, IAsyncDisposable
+	{
+		public bool WasDisposed { get; private set; }
+
+		public async ValueTask DisposeAsync()
+		{
+			await Task.Delay(10);
+			WasDisposed = true;
+		}
+	}
+
 	private static Mock<IComponentFactory> CreateMockComponentFactory(Func<Type, bool> canCreate, Func<Type, IComponent> create)
 	{
 		var result = new Mock<IComponentFactory>(MockBehavior.Strict);
@@ -54,4 +122,3 @@ public class TestContextBaseTest : TestContext
 	}
 }
 #endif
-